refactor(l5): move dream-light and dream-rem from src/ into server/maintenance (issue 864)
L5 of _plans/server-hardening-ladder.md. Neither module reads process.env, so
both move as-is; DreamEngine dry-run behavior is untouched.

Moved:
- src/dream-light.ts -> server/maintenance/dream-light.ts
- src/dream-rem.ts   -> server/maintenance/dream-rem.ts

Split by responsibility to meet the server/ function rules (no behavior change):
- server/maintenance/dream-light-occurrences.ts -- SweepRow plus the per-turn
  content_occurrences upsert lifted out of runLightSweep, which also removes
  the four-deep nesting the sweep loop carried.
- server/maintenance/dream-rem-grading.ts -- the candidate row shape, the row ->
  RemCandidate widening, the traced grade call, and the machine_grade UPDATE
  lifted out of runRemGrading.

Also hoisted the runRemPass argument assembly out of the dream.rem handler body
into buildRemPassDeps, and replaced the one non-null assertion in
readLightQueueDepth with an explicit guard on a query that always returns one row.

Shims (M2): src/dream-light.ts and src/dream-rem.ts now re-export from
server/maintenance so every legacy src/ and scripts/ importer is untouched.
Repointed only the server/ importer, server/domain/distill-window.ts.
Imports of already-moved modules (background-tracing, candidate-dedupe) name
their server/ paths directly; imports of still-in-src modules
(maintenance-queue, tools/ingest-raw-turn) use ../../src/ per M3.

Tests left in src/ per the M1 exception, all green through the shim:
src/dream-light.test.ts, src/dream-rem.test.ts, src/rem-terra-grader.test.ts,
src/distill-window.test.ts, src/maintenance-bootstrap.test.ts.

Runtime closure of src/ under server/main.ts: 46 -> 44.
